refactor: menugroup 패키지 이동
