+++ src/main/java/kitchenpos/domain/Menu.java
@@ -16,6 +16,7 @@ import javax.persistence.OneToMany;
 import javax.persistence.Table;
 import kitchenpos.common.domain.Name;
 import kitchenpos.common.domain.Price;
+import kitchenpos.menugroup.domain.MenuGroup;
 
 @Entity
 @Table(name = "menu")
