+++ src/main/java/kitchenpos/menugroup/application/MenuGroupService.java
@@ -1,8 +1,8 @@
-package kitchenpos.application;
+package kitchenpos.menugroup.application;
 
-import kitchenpos.domain.MenuGroupRepository;
-import kitchenpos.dto.MenuGroupRequest;
-import kitchenpos.dto.MenuGroupResponse;
+import kitchenpos.menugroup.domain.MenuGroupRepository;
+import kitchenpos.menugroup.dto.MenuGroupRequest;
+import kitchenpos.menugroup.dto.MenuGroupResponse;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
